Update to automerge@0.5.0, implement marks
There are some mechanical changes to the API:

* Allow negative `delete` arguments for `Doc.splice` and
  `Doc.spliceText`
* Add `Doc.mark`, `Doc.marks` and `Doc.marksAt`
* Add the `Mark` and `ExpandMark` types
* Add PatchAction.Conflict and PatchAction.Marks
* Add a `marks` field to `PatchAction.SpliceText` and
  `PatchAction.Insert`

One slightly more significant change is the addition of `unsafe impl
Send{}` and `unsafe impl Sync{}` on the `Doc` we export to uniffi. This
is required because currently `automerge::Automerge` is not `Send` or
`Sync` due to the internal use of an `Rc`. The `unsafe` impls are okay
because we guarantee that all accesses of the document on the Swift side
are from a single thread. In future we should be able to remove the `Rc`
from automerge to make it `Send` again and remove these unsafe impls.